requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
slowapi==0.1.9
llama-index==0.12.5
llama-index-vector-stores-chroma==0.4.1
llama-index-embeddings-huggingface==0.5.2
sentence-transformers==3.3.1
chromadb==0.6.3
python-multipart==0.0.20
pypdf==5.1.0
aiofiles==24.1.0
pydantic-settings==2.7.0
